requirements.txt
asgiref==3.7.2
boto3==1.28.64
botocore==1.31.64
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
django-storages==1.14.2
djangorestframework==3.14.0
google-api-core==2.12.0
google-auth==2.23.3
google-cloud-core==2.3.3
google-cloud-datastore==2.18.0
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.0
grpcio-status==1.59.0
idna==3.4
jmespath==1.0.1
proto-plus==1.22.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.6
gunicorn==21.2.0
whitenoise==5.3.0

